improv: better and more performant falling out of aether
