+++ src/main/java/com/aetherteam/aether/mixin/mixins/common/EntityMixin.java
@@ -1,0 +1,105 @@
+package com.aetherteam.aether.mixin.mixins.common;
+
+import com.aetherteam.aether.AetherConfig;
+import com.aetherteam.aether.AetherTags;
+import com.aetherteam.aether.block.portal.AetherPortalForcer;
+import com.aetherteam.aether.capability.item.DroppedItem;
+import com.aetherteam.aether.event.hooks.DimensionHooks;
+import com.aetherteam.aether.item.combat.abilities.armor.PhoenixArmor;
+import com.aetherteam.aether.network.AetherPacketHandler;
+import com.aetherteam.aether.network.packet.clientbound.SetVehiclePacket;
+import com.aetherteam.aether.world.LevelUtil;
+import com.aetherteam.nitrogen.network.PacketRelay;
+import net.minecraft.server.MinecraftServer;
+import net.minecraft.server.level.ServerLevel;
+import net.minecraft.server.level.ServerPlayer;
+import net.minecraft.world.entity.Entity;
+import net.minecraft.world.entity.LivingEntity;
+import net.minecraft.world.entity.Saddleable;
+import net.minecraft.world.entity.item.ItemEntity;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.entity.projectile.Projectile;
+import net.minecraft.world.level.Level;
+import net.minecraftforge.common.util.LazyOptional;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+
+import javax.annotation.Nullable;
+import java.util.List;
+
+@Mixin(Entity.class)
+public class EntityMixin {
+    /**
+     * Handles entities falling out of the Aether. If an entity is not a player, vehicle, or tracked item, it is removed.
+     *
+     * @param ci The {@link CallbackInfo} for the void method return.
+     * @see PhoenixArmor#boostVerticalLavaSwimming(LivingEntity)
+     */
+    @Inject(at = @At(value = "TAIL"), method = "tick()V")
+    private void travel(CallbackInfo ci) {
+        Entity entity = (Entity) (Object) this;
+        Level level = entity.level();
+        if (level instanceof ServerLevel serverLevel) {
+            if (!AetherConfig.SERVER.disable_falling_to_overworld.get()) {
+                if (serverLevel.dimension() == LevelUtil.destinationDimension()) {
+                    if (entity.getY() <= serverLevel.getMinBuildHeight() && !entity.isPassenger() && serverLevel.getBiome(entity.blockPosition()).is(AetherTags.Biomes.FALL_TO_OVERWORLD)) {
+                        if (entity instanceof Player || entity.isVehicle() || (entity instanceof Saddleable) && ((Saddleable) entity).isSaddled()) { // Checks if an entity is a player or a vehicle of a player.
+                            entityFell(entity);
+                        } else if (entity instanceof Projectile projectile && projectile.getOwner() instanceof Player) {
+                            entityFell(projectile);
+                        } else if (entity instanceof ItemEntity itemEntity) {
+                            LazyOptional<DroppedItem> droppedItem = DroppedItem.get(itemEntity);
+                            if (droppedItem.isPresent() && droppedItem.resolve().isPresent()) {
+                                if (itemEntity.getOwner() instanceof Player || droppedItem.resolve().get().getOwner() instanceof Player) { // Checks if an entity is an item that was dropped by a player.
+                                    entityFell(entity);
+                                }
+                            }
+                        }
+                    }
+                }
+            }
+        }
+    }
+
+    /**
+     * Code to handle falling out of the Aether with all passengers intact.
+     *
+     * @param entity The {@link Entity}
+     */
+    @Nullable
+    private static Entity entityFell(Entity entity) {
+        Level serverLevel = entity.level();
+        MinecraftServer minecraftserver = serverLevel.getServer();
+        if (minecraftserver != null) {
+            ServerLevel destination = minecraftserver.getLevel(LevelUtil.returnDimension());
+            if (destination != null && LevelUtil.returnDimension() != LevelUtil.destinationDimension()) {
+                List<Entity> passengers = entity.getPassengers();
+                serverLevel.getProfiler().push("aether_fall");
+                entity.setPortalCooldown();
+                Entity target = entity.changeDimension(destination, new AetherPortalForcer(destination, false));
+                serverLevel.getProfiler().pop();
+                // Check for passengers.
+                if (target != null) {
+                    for (Entity passenger : passengers) {
+                        passenger.stopRiding();
+                        Entity nextPassenger = entityFell(passenger);
+                        if (nextPassenger != null) {
+                            nextPassenger.startRiding(target);
+                            if (target instanceof ServerPlayer serverPlayer) { // Fixes a desync between the server and client.
+                                PacketRelay.sendToPlayer(AetherPacketHandler.INSTANCE, new SetVehiclePacket(nextPassenger.getId(), target.getId()), serverPlayer);
+                            }
+                        }
+                    }
+                    if (target instanceof ServerPlayer) {
+                        DimensionHooks.teleportationTimer = 500; // Sets a timer marking that the player teleported from falling out of the Aether.
+                    }
+                }
+                return target;
+            }
+        }
+        return null;
+    }
+}
+
